+++ IMNAT.School.Services/IMNAT.School.Services.csproj
@@ -5,6 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.5" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\Db_Context\Db_Context.csproj" />
     <ProjectReference Include="..\IMNAT.School.Repositories\IMNAT.School.Repositories.csproj" />
   </ItemGroup>
