--- IMNAT.School.Services/Services/IStudentManagement.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace IMNAT.School.Services.Services
-{
-    public interface IStudentManagement
-    {
-       void CreateStudent(string studentName, string email);
-
-    }
-}
